requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.6.1
pydantic-settings>=2.1.0
loguru>=0.7.2
sqlalchemy>=2.0.25
asyncpg>=0.29.0
redis>=5.0.1
alembic>=1.13.1
chromadb>=0.4.22
yfinance>=0.2.36
google-genai>=1.0.0
streamlit>=1.31.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
mypy>=1.3.0
httpx>=0.26.0
pandas>=2.2.0
apscheduler>=3.10.4
feedparser>=6.0.11
beautifulsoup4>=4.12.3
lxml>=5.1.0

tqdm>=4.66.0
